SQLAgent/requirements.txt
langchain>=0.2.0
langchain-google-genai>=0.1.2
google-generativeai>=0.8.3
langchain-community>=0.2.0
SQLAlchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.112.0
uvicorn>=0.30.0
requests>=2.32.0
typer>=0.12.0
PyMySQL>=1.1.0